Fix specs which couldn't handle the BOM in data
Normally, a BOM is only present in a file, not in the data read from the
file. But since some of the specs take shortcuts and assert on the
exporter's output rather than writing to a file and reading the file and
asserting on that, we need to manually remove the BOM in these specs.

Refs #108
Refs hitobito/hitobito#1993
Refs hitobito/hitobito#1960